// src/uni_modules/visual-components/components/visual-section/interface.ts
export interface VisualSectionProps {
  title?: string
  description?: string
  titleSize?: 'xs' | 'sm' | 'md' | 'lg' | 'base'
  titleColor?: string
  bold?: boolean
  descriptionSize?: 'xs' | 'sm' | 'md' | 'lg' | 'base'
  descriptionColor?: string
  showLine?: boolean
  lineColor?: string
  lineWidth?: string
  icon?: string
  iconColor?: string
}